refactor: 코스 필터의 거리 범위 슬라이더 로직 변경

+++ android/app/src/main/java/io/coursepick/coursepick/presentation/course/CoursesViewModel.kt
@@ -425,7 +425,7 @@ class CoursesViewModel
                 }
 
                 is CourseFilterAction.UpdateLengthRange -> {
-                    updateLengthRange(action.start, action.end)
+                    updateLengthRange(action.start.toDouble(), action.end.toDouble())
                 }
 
                 is CourseFilterAction.ToggleDifficulty -> {
